refactor(screens): reorder Main, Staking
