--- screens/MainScreen/index.ts
@@ -1,3 +1,0 @@
-import MainScreen from "./MainScreen";
-
-export default MainScreen;
