--- screens/StackingScreen/index.ts
@@ -1,3 +1,0 @@
-import StackingsScreen from "./StackingScreen";
-
-export default StackingsScreen;
