+++ screens/index.ts
@@ -5,4 +5,5 @@ export { default as ModalScreen } from "./ModalScreen";
 export { default as NotFoundScreen } from "./NotFoundScreen";
 export { default as TabOneScreen } from "./TabOneScreen";
 export { default as TabTwoScreen } from "./TabTwoScreen";
-export { default as MainScreen } from "./MainScreen/MainScreen";
+export { default as Main } from "./MainScreen";
+export { default as Stacking } from "./StackingScreen";
